+++ tests/consensus_tests/utils.py
@@ -9,6 +9,7 @@ import socket
 from contextlib import closing
 from pathlib import Path
 import pytest
+from .assertions import assert_http_ok
 
 # Tracks processes that need to be killed at the end of the test
 processes = []
@@ -43,15 +44,6 @@ def get_env(p2p_port: int, grpc_port: int, http_port: int) -> dict[str, str]:
 
 def get_uri(port: int) -> str:
     return f"http://127.0.0.1:{port}"
-
-
-def assert_http_ok(response: requests.Response):
-    if response.status_code != 200:
-        if not response.content:
-            raise Exception(f"Http request failed with status {response.status_code} and no content")
-        else:
-            raise Exception(
-                f"Http request failed with status {response.status_code} and contents:\n{response.json()}")
 
 
 def assert_project_root():
@@ -333,7 +325,7 @@ def wait_collection_on_all_peers(collection_name: str, peer_api_uris: [str], max
         exists = True
         for url in peer_api_uris:
             r = requests.get(f"{url}/collections")
-            assert r.status_code == 200
+            assert_http_ok(r)
             collections = r.json()["result"]["collections"]
             exists &= any(collection["name"] == collection_name for collection in collections)
         if exists:
